Breakup setup like Bell System
Remove get from setup
Move import out of setup
Breaking change v3.0.0
Print Minecraft version in tests

+++ src/mc_import.sh
@@ -3,32 +3,42 @@
 # Exit if error
 set -e
 jars_dir=~mc/java_jars
-syntax='Usage: mc_setup.sh INSTANCE'
+syntax='Usage: mc_import.sh [OPTION]... SERVER_DIR INSTANCE'
+update=true
 
-args=$(getopt -l help -o h -- "$@")
+args=$(getopt -l help,no-update -o hn -- "$@")
 eval set -- "$args"
 while [ "$1" != -- ]; do
 	case $1 in
 	--help|-h)
 		echo "$syntax"
-		echo 'Make new Minecraft Java Edition server in ~mc/java/INSTANCE.'
+		echo 'Import Minecraft Java Edition server to ~mc/java/INSTANCE.'
+		echo
+		echo Mandatory arguments to long options are mandatory for short options too.
+		echo "-n, --no-update  don't update minecraft java edition server"
 		exit
+		;;
+	--no-update|-n)
+		update=false
+		shift
 		;;
 	esac
 done
 shift
 
-if [ "$#" -lt 1 ]; then
+if [ "$#" -lt 2 ]; then
 	>&2 echo Not enough arguments
 	>&2 echo "$syntax"
 	exit 1
-elif [ "$#" -gt 1 ]; then
+elif [ "$#" -gt 2 ]; then
 	>&2 echo Too much arguments
 	>&2 echo "$syntax"
 	exit 1
 fi
 
-instance=$1
+import=$(realpath -- "$1")
+
+instance=$2
 if [ "$instance" != "$(systemd-escape -- "$instance")" ]; then
 	>&2 echo INSTANCE should be indentical to systemd-escape INSTANCE
 	exit 1
@@ -53,13 +63,29 @@ fi
 
 mkdir -p ~mc/java
 chown mc:mc ~mc/java
+
+echo "Enter Y if you stopped the server to import"
+read -r input
+input=$(echo "$input" | tr '[:upper:]' '[:lower:]')
+if [ "$input" != y ]; then
+	>&2 echo "$input != y"
+	exit 1
+fi
+
 trap 'rm -rf "$server_dir"' ERR
-mkdir "$server_dir"
-cp "$minecraft_jar" "$server_dir/server.jar"
-cd "$server_dir"
-# Minecraft Java Edition makes eula.txt on first run
-java -jar server.jar nogui || true
+cp -r "$import" "$server_dir"
+# Convert DOS line endings to UNIX line endings
+for file in "$server_dir"/*.{json,properties}; do
+	if [ -f "$file" ]; then
+		sed -i 's/\r$//' "$file"
+	fi
+done
 echo java -jar server.jar nogui > "$server_dir/start.bat"
 chmod +x "$server_dir/start.bat"
 chown -R mc:mc "$server_dir"
+if [ "$update" = true ]; then
+	cp "$minecraft_jar" "$server_dir/server.jar"
+fi
+trap - ERR
+rm -r "$import"
 echo "@@@ Remember to edit $server_dir/server.properties @@@"
